-Change names in function and commands names to be better understandable: SkinningMatrix => RelativeTransform, transformMatrix => absoluteMatrix -get in xml all unknow parameters, same if it's always 0 (because in futur, it could be not 0 on certain games) -rename xml unknow_X => unk_0 to simplify Xml.
-better deals with error :
    -asset converted (avoid just crash without infomration, cases are a solution each),
    -remove some debug to add some visual informations for users.
    -remove gatChar(), to use the notifyError() system.

-add a test witch valid the fact ema and ean need first keyframe at 0 , last at duration -1, and all are ordered in time.

+++ LibXenoverse/LibXenoverse/EAN.h
@@ -85,7 +85,7 @@ class EANAnimation;
 class EANKeyframe
 {
 public:
-	unsigned int frame;
+	size_t frame;
 	float x;
 	float y;
 	float z;
